add `__fspath__` and allow `__fspath__` to drive `__hash__`.
Since the source of the data should be unique on any running machine, this means the absolute filepath should be safe enough to use as a hash.

NOTE: This could break if the FeatureClass changes machine context or a webpath is modified. This is an uncommon occurance though and will almost certainly not happen during the lifetime of a `FeatureClass` (ominous foreshadowing sounds)